extract es doc build func
